+++ encuestas_lib/__init__.py
@@ -18,5 +18,5 @@ Uso:
     >>> df = IngestPipeline(config).run()
 """
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"
 __author__ = "Pablo Manrique"